Run HRDetect via signature.tools.lib

Runs HRDetect as described in the signature.tools.lib repository.

Usage

hrdetect_run(
  nm,
  snvindel_vcf,
  sv_vcf,
  cnv_tsv,
  genome = "hg38",
  snvoutdir = tempdir(),
  sigsToUse = c(1, 2, 3, 5, 6, 8, 13, 17, 18, 20, 26, 30),
  outpath = NULL
)

Arguments

nm

Sample name.

snvindel_vcf

Path to VCF with SNVs and INDELs.

sv_vcf

Path to VCF with SVs.

cnv_tsv

Path to purple.cnv.somatic.tsv file.

genome

Human genome version (default: hg38. hg19 means GRCh37).

snvoutdir

Directory to output SNV signature analysis results.

sigsToUse

COSMIC SNV signatures to use.

outpath

File to write HRDetect predictions to on disk (should end in '.gz'). If not specified, results won't be written to disk.

Value

Tibble with sample name and HRD probability in first two columns.

Examples

snvindel_vcf <- system.file(
  "extdata/umccrise/snv/somatic-ensemble-PASS.vcf.gz",
  package = "gpgr"
)
sv_vcf <- system.file("extdata/umccrise/sv/manta.vcf.gz", package = "gpgr")
cnv_tsv <- system.file("extdata/purple/purple.cnv.somatic.tsv", package = "gpgr")
nm <- "SampleA"
genome <- "hg38"
snvoutdir <- tempdir()
(res <- hrdetect_run(nm, snvindel_vcf, sv_vcf, cnv_tsv, genome, snvoutdir))
#> Bootstrap Signature Fits loaded from file
#> [info HRDetect_pipeline] Single Nucleotide Variations
#> [info HRDetect_pipeline] Structural Variants (Rearrangements)
#> [info HRDetect_pipeline] Some samples in the input data_matrix do not have the exposures for SV3 and SV5, checking if the user supplied SV catalogues or BEDPE files for those samples.
#> [info HRDetect_pipeline] Rearrangement signatures exposures will be estiamated for the following samples: SampleA
#> [info HRDetect_pipeline] Running Signature fit with 100 bootstraps. Increase sparsity by removing exposures with 5% threshold of total mutations and 0.05 threshold of p-value, i.e. exposure of a signature in a sample is set to zero if the probability of having less than 5% of total mutations assigned to that signature is greather than 0.05.
#> [info HRDetect_pipeline] Deletions at Micro-homology (Indels)
#> [info HRDetect_pipeline] HRD-LOH index (CNV)
#> [info HRDetect_pipeline] Computing HRDetect score and feature contributions for the following samples: SampleA
#> [info HRDetect_pipeline] HRDetect pipeline completed!
#> # A tibble: 1 × 9
#>   sample Probability intercept del.mh.prop  SNV3    SV3   SV5 hrdloh_index  SNV8
#>   <chr>        <dbl>     <dbl>       <dbl> <dbl>  <dbl> <dbl>        <dbl> <dbl>
#> 1 Sampl…       0.806     -3.36        6.20  2.44 -0.877 -1.10        -1.95 0.086
# hrdetect_run(nm, snvindel_vcf, sv_vcf, cnv_tsv, genome, snvoutdir,
#              outpath = "nogit/hrdetect_results.json.gz")