Process PURPLE CNV germline File for UMCCRISE

Processes the purple.cnv.germline.tsv file. and selects columns of interest.

Usage

purple_cnv_germ_process(x)

Arguments

x

Path to purple.cnv.germline.tsv file.

Value

List with two elements:

tab: Tibble with more condensed columns.

descr: Description of tibble columns.

Examples

x <- system.file("extdata/purple/purple.cnv.germline.tsv", package = "gpgr")
(pp <- purple_cnv_germ_process(x))
#> $tab
#> # A tibble: 67 × 9
#>    Chr     Start    End    CN `CN Min+Maj` `Start/End Seg…` Method `BAF (count)`
#>    <fct>   <int>  <int> <dbl> <chr>        <chr>            <chr>  <chr>        
#>  1 chr1   7.51e6 7.51e6   0   0+0          NONE-UNKNOWN     GERML… 0.98 (0)     
#>  2 chr1   1.41e7 1.41e7   0.4 0+0.4        NONE-UNKNOWN     GERML… 0.98 (0)     
#>  3 chr1   1.58e7 1.58e7   0.4 0+0.3        NONE-UNKNOWN     GERML… 0.98 (0)     
#>  4 chr1   5.83e7 5.83e7   0.3 0+0.3        NONE-UNKNOWN     GERML… 0.98 (0)     
#>  5 chr1   6.16e7 6.16e7   0.1 0+0.1        NONE-UNKNOWN     GERML… 0.98 (0)     
#>  6 chr1   7.98e7 7.98e7   0   0+0          NONE-UNKNOWN     GERML… 0.98 (0)     
#>  7 chr1   8.59e7 8.59e7   0.2 0+0.2        NONE-UNKNOWN     GERML… 0.98 (0)     
#>  8 chr1   8.90e7 8.90e7   0.1 0+0.1        NONE-UNKNOWN     GERML… 0.98 (0)     
#>  9 chr1   1.05e8 1.05e8   0.1 0+0.1        NONE-UNKNOWN     GERML… 0.98 (0)     
#> 10 chr1   1.11e8 1.11e8   0   0+0          NONE-UNKNOWN     GERML… 0.98 (0)     
#> # … with 57 more rows, and 1 more variable: `GC (windowCount)` <chr>
#> 
#> $descr
#> # A tibble: 7 × 2
#>   Column               Description                                              
#>   <chr>                <chr>                                                    
#> 1 Chr/Start/End        Coordinates of copy number segment                       
#> 2 CN                   Fitted absolute copy number of segment adjusted for puri…
#> 3 CN Min+Maj           CopyNumber of minor + major allele adjusted for purity   
#> 4 Start/End SegSupport Type of SV support for the CN breakpoint at start/end of…
#> 5 Method               Method used to determine the CN of the region. Allowed v…
#> 6 BAF (count)          Tumor BAF after adjusted for purity and ploidy (Count of…
#> 7 GC (windowCount)     Proportion of segment that is G or C (Count of COBALT wi…
#>