Process PURPLE CNV Gene File for UMCCRISE

Processes the purple.cnv.gene.tsv file. Keeps genes that are in the UMCCR cancer gene list (v24.03.0) and selects columns of interest.

Usage

purple_cnv_som_gene_process(x, g = NULL)

Arguments

x

Path to purple.cnv.gene.tsv file.

g

Path to gene file containing at least three columns:

ensembl_gene_symbol: gene name (character).

tsgene: is this gene a tumor suppressor (TRUE/FALSE).

oncogene: is this gene an oncogene (TRUE/FALSE).

Value

List with two elements:

tab: Tibble filtered to genes found in g.

descr: Description of tibble columns.

Examples

x <- system.file("extdata/purple/purple.cnv.gene.tsv", package = "gpgr")
g <- system.file("extdata/ref/somatic_panel-v24.03.0.tsv", package = "gpgr")
(pp <- purple_cnv_som_gene_process(x, g))
#> $tab
#> # A tibble: 5 × 14
#>   gene   minCN maxCN chrom   start     end chrBand onco_or_ts transcriptID   
#>   <chr>  <dbl> <dbl> <fct>   <int>   <int> <chr>   <chr>      <chr>          
#> 1 SDHA    5.66  5.66 chr5   218303  257082 p15.33  tsgene     ENST00000264932
#> 2 DUSP22  4.66  4.66 chr6   291630  351355 p25.3   tsgene     ENST00000419235
#> 3 IRF4    4.66  4.66 chr6   391752  411443 p25.3   oncogene   ENST00000380956
#> 4 KDM5A   4.30  4.30 chr12  280057  389320 p13.33  oncogene   ENST00000399788
#> 5 CRLF2   3.73  3.73 chrX  1187549 1212723 p22.33  oncogene   ENST00000400841
#> # ℹ 5 more variables: minMinorAlleleCN <dbl>, somReg <dbl>, minReg <dbl>,
#> #   minRegStartEnd <chr>, minRegSupportStartEndMethod <chr>
#> 
#> $descr
#> # A tibble: 11 × 2
#>    Column                      Description                                      
#>    <chr>                       <chr>                                            
#>  1 gene                        Name of gene                                     
#>  2 minCN/maxCN                 Min/Max copy number found in gene exons          
#>  3 chrom/start/end             Chromosome/start/end location of gene transcript 
#>  4 chrBand                     Chromosome band of the gene                      
#>  5 onco_or_ts                  oncogene ('oncogene'), tumor suppressor ('tsgene…
#>  6 transcriptID                Ensembl transcript ID (dot version)              
#>  7 minMinorAlleleCN            Minimum allele ploidy found over the gene exons …
#>  8 somReg (somaticRegions)     Count of somatic copy number regions this gene s…
#>  9 minReg (minRegions)         Number of somatic regions inside the gene that s…
#> 10 minRegStartEnd              Start/End base of the copy number region overlap…
#> 11 minRegSupportStartEndMethod Start/end support of the CN region overlapping t…
#>