Process PURPLE CNV Somatic File for UMCCRISE

Processes the purple.cnv.somatic.tsv file. and selects columns of interest.

purple_cnv_som_process(x)

Arguments

x

Path to purple.cnv.somatic.tsv file.

Value

List with two elements:

tab: Tibble with more condensed columns.

descr: Description of tibble columns.

Examples

x <- system.file("extdata/purple/purple.cnv.somatic.tsv", package = "gpgr")
(pp <- purple_cnv_som_process(x))
#> $tab
#> # A tibble: 59 × 9
#>    Chr       Start       End    CN `CN Min+Maj` Start/E…¹ Method BAF (…² GC (w…³
#>    <fct>     <int>     <int> <dbl> <chr>        <chr>     <chr>  <chr>   <chr>  
#>  1 chr1          1 123605522   1   0+1          TELOMERE… BAF_W… 0.98 (… 0.42 (…
#>  2 chr1  123605523 200044314   2   1+1          CENTROME… BAF_W… 0.5 (1… 0.4 (4…
#>  3 chr1  200044315 200044570   2.8 1+1.8        DUP-DUP   STRUC… 0.65 (… 0 (0)  
#>  4 chr1  200044571 248956422   2   1+1          DUP-TELO… BAF_W… 0.5 (1… 0.42 (…
#>  5 chr2          1  93139350   2   1+1          TELOMERE… BAF_W… 0.51 (… 0.41 (…
#>  6 chr2   93139351 219955359   2   1+1          CENTROME… BAF_W… 0.5 (1… 0.39 (…
#>  7 chr2  219955360 225225069   1   0+1          BND-BND   BAF_W… 0.98 (… 0.4 (5…
#>  8 chr2  225225070 242193529   2   1+1          BND-TELO… BAF_W… 0.51 (… 0.44 (…
#>  9 chr3          1  92214015   1   0+1          TELOMERE… BAF_W… 0.98 (… 0.4 (8…
#> 10 chr3   92214016 198295559   2   1+1          CENTROME… BAF_W… 0.5 (2… 0.39 (…
#> # … with 49 more rows, and abbreviated variable names ¹​`Start/End SegSupport`,
#> #   ²​`BAF (count)`, ³​`GC (windowCount)`
#> 
#> $descr
#> # A tibble: 7 × 2
#>   Column               Description                                              
#>   <chr>                <chr>                                                    
#> 1 Chr/Start/End        Coordinates of copy number segment                       
#> 2 CN                   Fitted absolute copy number of segment adjusted for puri…
#> 3 CN Min+Maj           CopyNumber of minor + major allele adjusted for purity   
#> 4 Start/End SegSupport Type of SV support for the CN breakpoint at start/end of…
#> 5 Method               Method used to determine the CN of the region. Allowed v…
#> 6 BAF (count)          Tumor BAF after adjusted for purity and ploidy (Count of…
#> 7 GC (windowCount)     Proportion of segment that is G or C (Count of COBALT wi…
#>