Read PURPLE Purity file

Reads the purple.purity.tsv file containing a summary of the purity fit.

Usage

purple_purity_read(x)

Arguments

x

Path to the purple.purity.tsv file.

Value

The input file as a tibble and a summarised tibble with a description of each metric.

Examples

x <- system.file("extdata/purple/purple.purity.tsv", package = "gpgr")
(p <- purple_purity_read(x))
#> $raw
#> # A tibble: 24 × 2
#>    column               value 
#>    <chr>                <chr> 
#>  1 purity               0.75  
#>  2 normFactor           1.05  
#>  3 score                0.26  
#>  4 diploidProportion    0.86  
#>  5 ploidy               1.87  
#>  6 gender               MALE  
#>  7 status               NORMAL
#>  8 polyclonalProportion 0     
#>  9 minPurity            0.71  
#> 10 maxPurity            0.78  
#> # … with 14 more rows
#> 
#> $summary
#> # A tibble: 10 × 4
#>        n variable        value            details                               
#>    <dbl> <chr>           <glue>           <chr>                                 
#>  1     2 Purity          0.75 (0.71-0.78) Purity of tumor in the sample (and mi…
#>  2     3 Ploidy          1.87 (1.86-1.88) Average ploidy of tumor sample after …
#>  3     4 Gender          MALE             Gender as inferred by AMBER/COBALT.   
#>  4     7 WGD             FALSE            Whole genome duplication (more than 1…
#>  5     8 MSI (indels/Mb) MSS (0)          MSI status (MSI, MSS or UNKNOWN if so…
#>  6     9 PolyclonalProp  0                Proportion of CN regions that are mor…
#>  7    10 DiploidyProp    0.86 (0.86-0.86) Proportion of CN regions that have 1 …
#>  8    11 TMB             0 (LOW)          Tumor mutational burden (# PASS varia…
#>  9    12 TML             0 (LOW)          Tumor mutational load (# of missense …
#> 10    13 TMB-SV          713              # of non inferred, non single passing…
#>