Read PURPLE Somatic SNV VCF

Reads the purple.somatic.vcf.gz file.

Usage

purple_snv_vcf_read(x)

Arguments

x

Path to the purple.somatic.vcf.gz file.

Value

A list with the input file as a tibble (with specific INFO fields) and a tibble with a description of each INFO field.

Examples

x <- system.file("extdata/purple/purple.somatic.vcf.gz", package = "gpgr")
(snv <- purple_snv_vcf_read(x))
#> $data
#> # A tibble: 423 × 13
#>    CHROM    POS AF    PURPLE_AF PURPLE_CN PURPLE_GERMLINE PURPLE_MACN PURPLE_VCN
#>    <chr>  <int> <chr>     <dbl>     <dbl> <chr>                 <dbl>      <dbl>
#>  1 chr1  3.09e6 0.14…     0.173         2 NOISE                 0.959      0.346
#>  2 chr1  3.29e6 0.24…     0.286         2 NOISE                 0.959      0.572
#>  3 chr1  3.40e6 0.11…     0.136         2 NOISE                 0.959      0.272
#>  4 chr1  6.96e6 0.5       0.581         2 NOISE                 0.959      1.16 
#>  5 chr1  6.98e6 0.5       0.581         2 NOISE                 0.959      1.16 
#>  6 chr1  7.10e6 0.51…     0.596         2 NOISE                 0.959      1.19 
#>  7 chr1  7.19e6 0.40…     0.476         2 NOISE                 0.959      0.951
#>  8 chr1  7.19e6 0.48…     0.568         2 NOISE                 0.959      1.14 
#>  9 chr1  7.19e6 0.63…     0.741         2 NOISE                 0.959      1.48 
#> 10 chr1  7.29e6 0.54…     0.630         2 NOISE                 0.959      1.26 
#> # ℹ 413 more rows
#> # ℹ 5 more variables: HMF_HOTSPOT <lgl>, KT <chr>, MH <chr>, SUBCL <dbl>,
#> #   TNC <chr>
#> 
#> $description
#> # A tibble: 11 × 2
#>    ID              Description                                                  
#>    <chr>           <chr>                                                        
#>  1 AF              Allele Frequency, for each ALT allele, in the same order as …
#>  2 HMF_HOTSPOT     calculated by flag of overlapping values in field HMF from s…
#>  3 KT              Forward/reverse kataegis id                                  
#>  4 MH              Microhomology                                                
#>  5 PURPLE_AF       Purity adjusted allelic frequency of variant                 
#>  6 PURPLE_CN       Purity adjusted copy number surrounding variant location     
#>  7 PURPLE_GERMLINE Germline classification surrounding variant location         
#>  8 PURPLE_MACN     Purity adjusted minor allele ploidy surrounding variant loca…
#>  9 PURPLE_VCN      Purity adjusted ploidy of variant                            
#> 10 SUBCL           Non-zero subclonal likelihood                                
#> 11 TNC             Tri-nucleotide context                                       
#>