Process Structural Variants

Processes the Manta TSV file output by umccrise.

Usage

process_sv(x)

Arguments

x

Path to manta.tsv output by umccrise.

Value

A list with melted/unmelted tibbles (these are NULL if TSV file was empty).

Examples

x <- system.file("extdata/umccrise/sv/manta.tsv", package = "gpgr")
(sv <- process_sv(x))
#> $unmelted
#> # A tibble: 58 × 19
#>    vcfnum  nann TierTop Start          End   Type  BND_ID BND_mate SR_alt PR_alt
#>    <chr>  <dbl> <chr>   <chr>          <chr> <chr> <chr>  <chr>     <int>  <int>
#>  1 16         1 4       7:77,167,955   7:10… BND   06     B            NA      9
#>  2 17         1 4       7:102,702,580  7:77… BND   06     A            NA      9
#>  3 20         1 3       8:70,067,891   8:70… BND   01     B            28     43
#>  4 21         2 3       8:70,106,277   8:70… BND   01     A            28     43
#>  5 28         1 4       10:101,127,983 14:2… BND   02     A            22     45
#>  6 29         1 4       10:101,127,987 14:2… BND   03     A            NA     59
#>  7 31         1 4       11:48,321,158  11:5… BND   05     B            NA     11
#>  8 32         1 4       11:50,058,044  11:4… BND   05     A            NA     11
#>  9 37         2 2       14:22,430,592  10:1… BND   03     B            NA     59
#> 10 38         7 4       14:22,430,590  14:2… BND   04     B            40     75
#> # … with 48 more rows, and 9 more variables: SR_PR_sum <int>, SR_PR_ref <chr>,
#> #   Ploidy <dbl>, AF_PURPLE <chr>, AF_BPI <chr>, CNC <chr>, CN <chr>,
#> #   SScore <int>, annotation <chr>
#> 
#> $melted
#> # A tibble: 1,056 × 28
#>    vcfnum  nann TierTop Start         End    Type  BND_ID BND_mate SR_alt PR_alt
#>    <chr>  <dbl> <chr>   <chr>         <chr>  <chr> <chr>  <chr>     <int>  <int>
#>  1 02         2 1       1:243,760,053 1:244… DEL   NA     NA           34     66
#>  2 06         2 2       4:108,131,468 4:108… DEL   NA     NA           39     71
#>  3 22         5 2       8:70,106,311  8:70,… DEL   NA     NA           25     47
#>  4 37         2 2       14:22,430,592 10:10… BND   03     B            NA     59
#>  5 02         2 1       1:243,760,053 1:244… DEL   NA     NA           34     66
#>  6 22         5 2       8:70,106,311  8:70,… DEL   NA     NA           25     47
#>  7 22         5 2       8:70,106,311  8:70,… DEL   NA     NA           25     47
#>  8 01        80 3       1:3,660,719   1:6,6… DEL   NA     NA           27     68
#>  9 30        65 3       11:9,681,908  11:12… DEL   NA     NA           29     71
#> 10 33       333 3       12:8,384,896  12:15… DEL   NA     NA           29     69
#> # … with 1,046 more rows, and 18 more variables: SR_PR_sum <int>,
#> #   SR_PR_ref <chr>, Ploidy <dbl>, AF_PURPLE <chr>, AF_BPI <chr>, CNC <chr>,
#> #   CN <chr>, SScore <int>, Event <chr>, Effect <chr>, Genes <chr>,
#> #   Transcript <chr>, Detail <chr>, Tier <chr>, ntrx <dbl>, ngen <dbl>,
#> #   neff <dbl>, `Tier (Top)` <glue>
#> 
#> $tsv_descr
#> # A tibble: 23 × 3
#>    Column             Description                                          Type 
#>    <chr>              <chr>                                                <chr>
#>  1 caller             Manta SV caller                                      c    
#>  2 sample             Tumor sample name                                    c    
#>  3 chrom              CHROM column in VCF                                  c    
#>  4 start              POS column in VCF                                    i    
#>  5 end                INFO/END: End position of the variant described in … i    
#>  6 svtype             INFO/SVTYPE: Type of structural variant              c    
#>  7 split_read_support FORMAT/SR of tumor sample: Split reads for the ref … c    
#>  8 paired_support_PE  FORMAT/PE of tumor sample: ??                        c    
#>  9 paired_support_PR  FORMAT/PR of tumor sample: Spanning paired-read sup… c    
#> 10 AF_BPI             INFO/BPI_AF: AF at each breakpoint (so AF_BPI1,AF_B… c    
#> # … with 13 more rows
#> 
#> $col_descr
#> # A tibble: 30 × 2
#>    Column     Description                                                       
#>    <chr>      <chr>                                                             
#>  1 nrow       "Row number that connects variants between tables in same tab set…
#>  2 vcfnum     "Original event row number that connects variants to events."     
#>  3 TierTop    "Top priority of the event (from simple_sv_annotation: 1 highest,…
#>  4 Tier       "Priority of the specific event (from simple_sv_annotation: 1 hig…
#>  5 Tier (Top) "Priority of the specific event (Top tier of all event's annotati…
#>  6 Chr        "Chromosome."                                                     
#>  7 Start      "Start position as inferred by BPI. For PURPLE-inferred SVs we us…
#>  8 End        "End position. For BNDs = Chr:Start of the mate.Values are inferr…
#>  9 ID         "ID of BND from Manta. For PURPLE-inferred SVs this is PURPLE."   
#> 10 MATEID     "ID of BND mate from Manta."                                      
#> # … with 20 more rows
#>