Process PURPLE CNV Gene File for UMCCRISE

Processes the purple.cnv.gene.tsv file. Keeps genes that are in the UMCCR cancer gene list and selects columns of interest.

purple_cnv_som_gene_process(x, g = NULL)

Arguments

x

Path to purple.cnv.gene.tsv file.

g

Path to gene file containing at least three columns:

symbol: gene name (character).

tumorsuppressor: is this gene a tumor suppressor (TRUE/FALSE).

oncogene: is this gene an oncogene (TRUE/FALSE).

Value

List with two elements:

tab: Tibble filtered to genes found in g.

descr: Description of tibble columns.

Examples

x <- system.file("extdata/purple/purple.cnv.gene.tsv", package = "gpgr")
g <- system.file("extdata/ref/umccr_cancer_genes_2019-03-20.tsv", package = "gpgr")
(pp <- purple_cnv_som_gene_process(x, g))
#> $tab
#> # A tibble: 13 × 15
#>    gene   minCN maxCN chrom  start    end chrBand onco_…¹ trans…² minMi…³ somReg
#>    <chr>  <dbl> <dbl> <fct>  <int>  <int> <chr>   <chr>   <chr>     <dbl>  <dbl>
#>  1 CRBN   1.03  1.03  chr3  3.15e6 3.18e6 p26.2   ""      ENST00…  0.0176      1
#>  2 SDHA   2.01  2.01  chr5  2.18e5 2.57e5 p15.33  "tsgen… ENST00…  0.994       1
#>  3 DUSP22 2.00  2.00  chr6  2.92e5 3.51e5 p25.3   ""      ENST00…  0.987       1
#>  4 IRF4   2.00  2.00  chr6  3.92e5 4.11e5 p25.3   "oncog… ENST00…  0.987       1
#>  5 FOXQ1  2.00  2.00  chr6  1.31e6 1.31e6 p25.3   ""      ENST00…  0.987       1
#>  6 DOCK8  1.99  1.99  chr9  2.15e5 4.65e5 p24.3   ""      ENST00…  0.989       1
#>  7 LARP4B 1.99  1.99  chr10 8.07e5 9.32e5 p15.3   "tsgen… ENST00…  0.982       1
#>  8 SIRT3  2.00  2.00  chr11 2.15e5 2.36e5 p15.5   "tsgen… ENST00…  0.989       1
#>  9 KDM5A  2.00  2.00  chr12 2.80e5 3.89e5 p13.33  "oncog… ENST00…  0.988       1
#> 10 ZMYM2  2.00  2.00  chr13 2.00e7 2.01e7 q12.11  ""      ENST00…  0.994       1
#> 11 TYMS   2.01  2.01  chr18 6.58e5 6.74e5 p11.32  ""      ENST00…  0.993       1
#> 12 CRLF2  0.997 0.997 chrX  1.19e6 1.21e6 p22.33  "oncog… ENST00…  0           1
#> 13 SRY    1.01  1.01  chrY  2.79e6 2.79e6 p11.2   ""      ENST00…  0           1
#> # … with 4 more variables: germDelReg <chr>, minReg <dbl>,
#> #   minRegStartEnd <chr>, minRegSupportStartEndMethod <chr>, and abbreviated
#> #   variable names ¹​onco_or_ts, ²​transcriptID, ³​minMinorAlleleCN
#> 
#> $descr
#> # A tibble: 12 × 2
#>    Column                                                                Descr…¹
#>    <chr>                                                                 <chr>  
#>  1 gene                                                                  Name o…
#>  2 minCN/maxCN                                                           Min/Ma…
#>  3 chrom/start/end                                                       Chromo…
#>  4 chrBand                                                               Chromo…
#>  5 onco_or_ts                                                            oncoge…
#>  6 transcriptID                                                          Ensemb…
#>  7 minMinorAlleleCN                                                      Minimu…
#>  8 somReg (somaticRegions)                                               Count …
#>  9 germDelReg (germlineHomDeletionRegions / germlineHetToHomDeletionReg… Number…
#> 10 minReg (minRegions)                                                   Number…
#> 11 minRegStartEnd                                                        Start/…
#> 12 minRegSupportStartEndMethod                                           Start/…
#> # … with abbreviated variable name ¹​Description
#>