Process PURPLE CNV Somatic File for UMCCRISE

Processes the purple.cnv.somatic.tsv file. and selects columns of interest.

Usage

purple_cnv_som_process(x)

Arguments

x

Path to purple.cnv.somatic.tsv file.

Value

List with two elements:

tab: Tibble with more condensed columns.

descr: Description of tibble columns.

Examples

x <- system.file("extdata/purple/purple.cnv.somatic.tsv", package = "gpgr")
(pp <- purple_cnv_som_process(x))
#> $tab
#> # A tibble: 59 × 9
#>    Chr     Start    End    CN `CN Min+Maj` `Start/End Seg…` Method `BAF (count)`
#>    <fct>   <int>  <int> <dbl> <chr>        <chr>            <chr>  <chr>        
#>  1 chr1   1   e0 1.24e8   1   0+1          TELOMERE-CENTRO… BAF_W… 0.98 (20830) 
#>  2 chr1   1.24e8 2.00e8   2   1+1          CENTROMERE-DUP   BAF_W… 0.5 (10239)  
#>  3 chr1   2.00e8 2.00e8   2.8 1+1.8        DUP-DUP          STRUC… 0.65 (0)     
#>  4 chr1   2.00e8 2.49e8   2   1+1          DUP-TELOMERE     BAF_W… 0.5 (10341)  
#>  5 chr2   1   e0 9.31e7   2   1+1          TELOMERE-CENTRO… BAF_W… 0.51 (19339) 
#>  6 chr2   9.31e7 2.20e8   2   1+1          CENTROMERE-BND   BAF_W… 0.5 (19047)  
#>  7 chr2   2.20e8 2.25e8   1   0+1          BND-BND          BAF_W… 0.98 (1284)  
#>  8 chr2   2.25e8 2.42e8   2   1+1          BND-TELOMERE     BAF_W… 0.51 (4236)  
#>  9 chr3   1   e0 9.22e7   1   0+1          TELOMERE-CENTRO… BAF_W… 0.98 (18629) 
#> 10 chr3   9.22e7 1.98e8   2   1+1          CENTROMERE-TELO… BAF_W… 0.5 (21114)  
#> # … with 49 more rows, and 1 more variable: `GC (windowCount)` <chr>
#> 
#> $descr
#> # A tibble: 7 × 2
#>   Column               Description                                              
#>   <chr>                <chr>                                                    
#> 1 Chr/Start/End        Coordinates of copy number segment                       
#> 2 CN                   Fitted absolute copy number of segment adjusted for puri…
#> 3 CN Min+Maj           CopyNumber of minor + major allele adjusted for purity   
#> 4 Start/End SegSupport Type of SV support for the CN breakpoint at start/end of…
#> 5 Method               Method used to determine the CN of the region. Allowed v…
#> 6 BAF (count)          Tumor BAF after adjusted for purity and ploidy (Count of…
#> 7 GC (windowCount)     Proportion of segment that is G or C (Count of COBALT wi…
#>