Read PURPLE Somatic SNV VCF

Reads the purple.somatic.vcf.gz file.

purple_snv_vcf_read(x)

Arguments

x

Path to the purple.somatic.vcf.gz file.

Value

A list with the input file as a tibble (with specific INFO fields) and a tibble with a description of each INFO field.

Examples

x <- system.file("extdata/purple/purple.somatic.vcf.gz", package = "gpgr")
(snv <- purple_snv_vcf_read(x))
#> $data
#> # A tibble: 423 × 13
#>    CHROM     POS AF     PURPLE_AF PURPLE…¹ PURPL…² PURPL…³ PURPL…⁴ HMF_H…⁵ KT   
#>    <chr>   <int> <chr>      <dbl>    <dbl> <chr>     <dbl>   <dbl> <lgl>   <chr>
#>  1 chr1  3086201 0.1486     0.173        2 NOISE     0.959   0.346 FALSE   NA   
#>  2 chr1  3294434 0.2462     0.286        2 NOISE     0.959   0.572 FALSE   NA   
#>  3 chr1  3400513 0.1169     0.136        2 NOISE     0.959   0.272 FALSE   NA   
#>  4 chr1  6957208 0.5        0.581        2 NOISE     0.959   1.16  FALSE   NA   
#>  5 chr1  6975198 0.5        0.581        2 NOISE     0.959   1.16  FALSE   NA   
#>  6 chr1  7100329 0.5122     0.596        2 NOISE     0.959   1.19  FALSE   NA   
#>  7 chr1  7185898 0.4091     0.476        2 NOISE     0.959   0.951 FALSE   NA   
#>  8 chr1  7186176 0.4884     0.568        2 NOISE     0.959   1.14  FALSE   NA   
#>  9 chr1  7186250 0.6375     0.741        2 NOISE     0.959   1.48  FALSE   NA   
#> 10 chr1  7290559 0.5422     0.630        2 NOISE     0.959   1.26  FALSE   NA   
#> # … with 413 more rows, 3 more variables: MH <chr>, SUBCL <dbl>, TNC <chr>, and
#> #   abbreviated variable names ¹​PURPLE_CN, ²​PURPLE_GERMLINE, ³​PURPLE_MACN,
#> #   ⁴​PURPLE_VCN, ⁵​HMF_HOTSPOT
#> 
#> $description
#> # A tibble: 11 × 2
#>    ID              Description                                                  
#>    <chr>           <chr>                                                        
#>  1 AF              Allele Frequency, for each ALT allele, in the same order as …
#>  2 HMF_HOTSPOT     calculated by flag of overlapping values in field HMF from s…
#>  3 KT              Forward/reverse kataegis id                                  
#>  4 MH              Microhomology                                                
#>  5 PURPLE_AF       Purity adjusted allelic frequency of variant                 
#>  6 PURPLE_CN       Purity adjusted copy number surrounding variant location     
#>  7 PURPLE_GERMLINE Germline classification surrounding variant location         
#>  8 PURPLE_MACN     Purity adjusted minor allele ploidy surrounding variant loca…
#>  9 PURPLE_VCN      Purity adjusted ploidy of variant                            
#> 10 SUBCL           Non-zero subclonal likelihood                                
#> 11 TNC             Tri-nucleotide context                                       
#>